pyproject.toml
[project]
name = "urban-platform"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]==5.6.2",
    "django==6.0.3",
    "django-celery-beat", # we temporarily install this from git because Django 6.0 support is not yet released
    "gunicorn==25.1.0",
    "pre-commit==4.5.1",
    "psycopg==3.3.3",
    "python-dotenv==1.2.2",
    "python-slugify==8.0.4",
    "redis==6.4.0",
    "sentry-sdk==2.54.0",
    "sib-api-v3-sdk==7.6.0",
    "tqdm==4.67.3",
    "wagtail==7.3.1",
]

[tool.uv.sources]
django-celery-beat = { git = "https://github.com/celery/django-celery-beat", rev = "5ca807e9fe446961073ee5d594fea0124ba0e6c0" }

[tool.mypy]
plugins = [
    "mypy_django_plugin.main"
]
strict_optional = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["wagtail.*", "modelcluster.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["*.tests.*"]
disallow_untyped_defs = false

[tool.django-stubs]
django_settings_module = "urban_platform.settings.dev"

[dependency-groups]
dev = [
    "celery-stubs>=0.1.3",
    "django-browser-reload>=1.21.0",
    "django-stubs[compatible-mypy]>=5.2.8",
    "factory-boy>=3.3.3",
    "mypy>=1.19.1",
    "playwright>=1.57.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-env>=1.2.0",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
    "ruff>=0.9.4",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "urban_platform.settings.test"
addopts = "--strict-markers --tb=short -m 'not e2e'"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
markers = [
    "e2e: marks tests as end-to-end tests (deselect with '-m \"not e2e\"')",
]

[tool.coverage.run]
relative_files = true
branch = true
source = [
    "about",
    "core",
    "home",
    "legal",
    "pedagogy",
    "publications",
    "search",
]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/tests.py",
    "*/factories/*",
    "*/factories.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
precision = 2

[tool.djlint]
profile = "django"
blank_line_after_tag = "load,extends"
close_void_tags = true
line_break_after_multiline_tag = true

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
ignore = [
    "F401" # use explicit re-export
]

[tool.ruff.lint.per-file-ignores]
"urban_platform/settings/*" = ["F403", "F405"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
